requirements.txt
yt-dlp
moviepy
google-generativeai
openai-whisper